Removed functionality
As of the previous update this template is not needed but is still called in some of the code so i just thought id leave it in and also if some one did want to make a customised template. added more medical items to med drop, as the resupply has med items in it so made it more worth while to call in the med also will help in adv med missions.